Refactor architecture violation tests and update package structure
