--- popcornguineapigplugin/src/test/kotlin/com/github/codandotv/popcorn/presentation/ext/TargetModuleExtTest.kt
@@ -1,7 +1,0 @@
-package com.github.codandotv.popcorn.presentation.ext
-
-import com.github.codandotv.popcorn.domain.metadata.InternalDependenciesMetadata
-import com.github.codandotv.popcorn.domain.metadata.TargetModule
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
